backend/requirements.txt
openai
fastapi
uvicorn
python-docx
tika
PyPDF2
python-dotenv